+++ components/frontend/src/source/SourceEntity.test.js
@@ -3,12 +3,13 @@ import { Table } from 'semantic-ui-react';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { SourceEntity } from './SourceEntity';
 
-function renderSourceEntity({ status = "unconfirmed", hide_ignored_entities = false }) {
+function renderSourceEntity({ status = "unconfirmed", status_end_date = "", hide_ignored_entities = false }) {
     return render(
         <Table>
             <Table.Body>
                 <SourceEntity
                     status={status}
+                    status_end_date={status_end_date}
                     hide_ignored_entities={hide_ignored_entities}
                     entity_attributes={[{key: "attr1"}, {key: "attr2", color: {bad: "warning"}}]}
                     entity={{ attr1: "good", attr2: "bad" }}
@@ -28,10 +29,20 @@ it('renders the unconfirmed status', () => {
 
 it('renders the fixed status', () => {
     renderSourceEntity({ status: "fixed" });
-    expect(screen.getAllByText(/Fixed/).length).toBe(1);
+    expect(screen.getAllByText(/Will be fixed/).length).toBe(1);
+})
+
+it('renders the status end date', () => {
+    renderSourceEntity({ status: "fixed", status_end_date: "3000-01-01" });
+    expect(screen.getAllByText(/Will be fixed \(status accepted until 3000-01-01\)/).length).toBe(1);
+})
+
+it('renders the status past end date', () => {
+    renderSourceEntity({ status: "fixed", status_end_date: "2000-01-01", hide_ignored_entities: true });
+    expect(screen.getAllByText(/Will be fixed \(status accepted until 2000-01-01\)/).length).toBe(1);
 })
 
 it('renders nothing if the status is to be ignored', () => {
     renderSourceEntity({ status: "fixed", hide_ignored_entities: true });
-    expect(screen.queryAllByText(/Fixed/).length).toBe(0);
+    expect(screen.queryAllByText(/Will be fixed/).length).toBe(0);
 })