+++ components/frontend/src/source/SourceEntityDetails.js
@@ -1,7 +1,8 @@
 import React from 'react';
-import { Grid, Header } from 'semantic-ui-react';
-import { TextInput } from '../fields/TextInput';
+import { Grid, Header, Icon, Popup } from 'semantic-ui-react';
+import { DateInput } from '../fields/DateInput';
 import { SingleChoiceInput } from '../fields/SingleChoiceInput';
+import { TextInput } from '../fields/TextInput';
 import { set_source_entity_attribute } from '../api/source';
 import { capitalize } from '../utils';
 import { source_entity_status_name as status_name } from './source_entity_status';
@@ -16,18 +17,18 @@ function entity_status_option(status, text, content, subheader) {
 function entity_status_options(entity_type) {
     return [
         entity_status_option('unconfirmed', status_name.unconfirmed, 'Unconfirm', `This ${entity_type} should be reviewed to decide what to do with it.`),
-        entity_status_option('fixed', status_name.fixed, "Resolve as fixed", `This ${entity_type} has been fixed and will disappear shortly.`),
+        entity_status_option('fixed', status_name.fixed, "Resolve as will be fixed", `This ${entity_type} will be fixed shortly and then disappear.`),
         entity_status_option('false_positive', status_name.false_positive, 'Resolve as false positive', `This ${entity_type} can be ignored because it's been incorrectly identified as ${entity_type}.`),
         entity_status_option('wont_fix', status_name.wont_fix, "Resolve as won't fix", `This ${entity_type} will not be fixed.`),
         entity_status_option('confirmed', status_name.confirmed, 'Confirm', `This ${entity_type} has been reviewed and should be dealt with.`),
     ]
 }
 
-export function SourceEntityDetails({ metric_uuid, source_uuid, entity, name, status, rationale, reload}) {
+export function SourceEntityDetails({ entity, metric_uuid, name, rationale, reload, status, status_end_date, source_uuid }) {
     return (
         <Grid stackable>
-            <Grid.Row columns={4}>
-                <Grid.Column width={4}>
+            <Grid.Row>
+                <Grid.Column width={3}>
                     <SingleChoiceInput
                         requiredPermissions={[EDIT_ENTITY_PERMISSION]}
                         label={`${capitalize(name)} status`}
@@ -37,11 +38,21 @@ export function SourceEntityDetails({ metric_uuid, source_uuid, entity, name, st
                         sort={false}
                     />
                 </Grid.Column>
-                <Grid.Column width={12}>
+                <Grid.Column width={3}>
+                    <DateInput
+                        requiredPermissions={[EDIT_ENTITY_PERMISSION]}
+                        label={<label>{`${capitalize(name)} status end date`} <Popup on={['hover', 'focus']} content={`Consider the status of this ${name} to be 'Unconfirmed' after the selected date.`} trigger={<Icon tabIndex="0" name="help circle" />} /></label>}
+                        placeholder="YYYY-MM-DD"
+                        set_value={(value) => set_source_entity_attribute(metric_uuid, source_uuid, entity.key, "status_end_date", value, reload)}
+                        value={status_end_date}
+                    />
+                </Grid.Column>
+                <Grid.Column width={10}>
                     <TextInput
                         requiredPermissions={[EDIT_ENTITY_PERMISSION]}
                         label="Rationale"
                         placeholder={`Rationale for ${name} status...`}
+                        rows={Math.min(5, rationale?.split("\n").length ?? 1)}
                         set_value={(value) => set_source_entity_attribute(metric_uuid, source_uuid, entity.key, "rationale", value, reload)}
                         value={rationale}
                     />