+++ components/frontend/src/source/SourceEntityDetails.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { fireEvent, render, screen } from '@testing-library/react';
+import { act, fireEvent, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { EDIT_ENTITY_PERMISSION, Permissions } from '../context/Permissions';
 import * as source from '../api/source';
@@ -32,6 +32,15 @@ it('changes the entity status', () => {
     fireEvent.click(screen.getByText(/Confirm/))
     expect(source.set_source_entity_attribute).toHaveBeenCalledWith(
         "metric_uuid", "source_uuid", "key", "status", "confirmed", reload
+    );
+})
+
+it('changes the entity status end date', async () => {
+    source.set_source_entity_attribute = jest.fn()
+    renderSourceEntityDetails()
+    await act(async () => fireEvent.change(screen.getByPlaceholderText(/YYYY-MM-DD/), {target: {value: '2222-01-01'}}))
+    expect(source.set_source_entity_attribute).toHaveBeenCalledWith(
+        "metric_uuid", "source_uuid", "key", "status_end_date", "2222-01-01", reload
     );
 })